+++ app.rb
@@ -1,5 +1,9 @@
 # frozen_string_literal: true
 
+Dir['./lib/mcm/**/*.rb'].sort.each do |file|
+  require file
+end
+
 configure do
   set :session_timeout, 8 * 3600 # seconds
